--- lastSubject/src/main/java/com/raiseTech/lastSubject/LastSubjectApplication.java
@@ -1,13 +1,0 @@
-package com.raiseTech.lastSubject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LastSubjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LastSubjectApplication.class, args);
-	}
-
-}
