--- lastSubject/src/main/java/com/raiseTech/lastSubject/ServletInitializer.java
@@ -1,13 +1,0 @@
-package com.raiseTech.lastSubject;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(LastSubjectApplication.class);
-	}
-
-}
