--- lastSubject/src/main/java/com/raiseTech/lastSubject/entity/MedicineInformation.java
@@ -1,6 +1,0 @@
-package com.raiseTech.lastSubject.entity;
-
-import java.sql.Timestamp;
-
-public record MedicineInformation(int id, int userId, Timestamp visitDate, String pharmacy, String medicine) {
-}
