--- lastSubject/src/main/java/com/raiseTech/lastSubject/entity/PatientInformation.java
@@ -1,6 +1,0 @@
-package com.raiseTech.lastSubject.entity;
-
-import java.sql.Date;
-
-public record PatientInformation(int id, String name, Date birthdate) {
-}
