--- lastSubject/src/main/java/com/raiseTech/lastSubject/exeption/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package com.raiseTech.lastSubject.exeption;
-
-public class ResourceNotFoundException extends RuntimeException {
-	public ResourceNotFoundException(String message) {
-		super(message);
-	}
-}
