--- lastSubject/src/main/java/com/raiseTech/lastSubject/request/PatientRequest.java
@@ -1,18 +1,0 @@
-package com.raiseTech.lastSubject.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PatientRequest {
-	private int id;
-	private String name;
-	private int birthdate;
-	public PatientRequest(String name, int birthdate) {
-		this.name = name;
-		this.birthdate = birthdate;
-	}
-	public PatientRequest() {
-	}
-}
