--- lastSubject/src/main/java/com/raiseTech/lastSubject/response/PatientsResponse.java
@@ -1,23 +1,0 @@
-package com.raiseTech.lastSubject.response;
-
-import com.raiseTech.lastSubject.entity.PatientInformation;
-import lombok.Getter;
-
-import java.sql.Date;
-
-@Getter
-public class PatientsResponse {
-	private final int id;
-	private final String name;
-	private final Date birthdate;
-	public PatientsResponse(PatientInformation patientInformation) {
-		this.id = patientInformation.id();
-		this.name = patientInformation.name();
-		this.birthdate = patientInformation.birthdate();
-	}
-	public PatientsResponse(int id, String name, Date birthdate) {
-		this.id = id;
-		this.name = name;
-		this.birthdate = birthdate;
-	}
-}
