+++ lastSubject/src/main/java/com/raisetech/lastsubject/LastSubjectApplication.java
@@ -1,0 +1,13 @@
+package com.raisetech.lastsubject;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LastSubjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LastSubjectApplication.class, args);
+	}
+
+}
