+++ lastSubject/src/main/java/com/raisetech/lastsubject/ServletInitializer.java
@@ -1,0 +1,13 @@
+package com.raisetech.lastsubject;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(LastSubjectApplication.class);
+	}
+
+}
