+++ lastSubject/src/main/java/com/raisetech/lastsubject/entity/MedicineInformation.java
@@ -1,0 +1,6 @@
+package com.raisetech.lastsubject.entity;
+
+import java.sql.Timestamp;
+
+public record MedicineInformation(int id, int userId, Timestamp visitDate, String pharmacy, String medicine) {
+}
