+++ lastSubject/src/main/java/com/raisetech/lastsubject/entity/PatientInformation.java
@@ -1,0 +1,6 @@
+package com.raisetech.lastsubject.entity;
+
+import java.sql.Date;
+
+public record PatientInformation(int id, String name, Date birthdate) {
+}
