+++ lastSubject/src/main/java/com/raisetech/lastsubject/exeption/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.raisetech.lastsubject.exeption;
+
+public class ResourceNotFoundException extends RuntimeException {
+	public ResourceNotFoundException(String message) {
+		super(message);
+	}
+}
