+++ lastSubject/src/main/resources/com/raisetech/lastsubject/mapper/MedicineNotebookMapper.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.raisetech.lastsubject.mapper.MedicineNotebookMapper">
+    <select id="findAll" parameterType="java.util.List"
+            resultType="com.raisetech.lastsubject.entity.PatientInformation">
+        select * from patient_informations
+    </select>
+
+    <select id="findByPatient" parameterType="java.util.List"
+            resultType="com.raisetech.lastsubject.entity.MedicineInformation">
+        select * from patient_medicine_informations where userId=(select id from patient_informations where name=#{name}
+        and
+        birthdate=#{birthdate})
+    </select>
+
+    <insert id="postPatient" useGeneratedKeys="true" keyProperty="id">
+        insert into patient_informations (name,birthdate) values (#{name},#{birthdate})
+        <selectKey resultType="int" keyProperty="id" order="AFTER">
+            select @@IDENTITY
+        </selectKey>
+    </insert>
+
+    <insert id="postMedicine" useGeneratedKeys="true" keyProperty="id">
+        insert into patient_medicine_informations(userId,visitDate,pharmacy,medicine) values
+        ((select id from patient_informations where name=#{name}
+        and
+        birthdate=#{birthdate}),now(),#{pharmacy},#{medicine})
+        <selectKey resultType="int" keyProperty="id" order="AFTER">
+            select @@IDENTITY
+        </selectKey>
+    </insert>
+</mapper>
