--- lastSubject/src/test/java/com/raiseTech/lastSubject/LastSubjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.raiseTech.lastSubject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LastSubjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
