+++ lastSubject/src/test/java/com/raisetech/lastsubject/LastSubjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.raisetech.lastsubject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LastSubjectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
